requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
email_validator
Werkzeug==2.3.8  
